+++ packages/widgetbook_cli/lib/src/cache/cache.dart
@@ -2,4 +2,7 @@ export 'addons_configs.dart';
 export 'cache_exception.dart';
 export 'cache_reader.dart';
 export 'cache_store.dart';
-export 'use_case_metadata.dart';
+export 'component_metadata.dart';
+export 'scenario_metadata.dart';
+export 'scenario_record.dart';
+export 'story_metadata.dart';