+++ packages/widgetbook_cli/CHANGELOG.md
@@ -1,3 +1,7 @@
+# Unreleased
+
+- **BREAKING**: Support Widgetbook v4 in `cloud build push` command. ([#1638](https://github.com/widgetbook/widgetbook/pull/1638))
+
 ## 3.10.0
 
 - **BREAKING**: Set minimum SDK version to 3.7.0. ([#1541](https://github.com/widgetbook/widgetbook/pull/1541))